+++ lib/src/plugins/importer.rs
@@ -2,7 +2,7 @@
 Importers allow users to (periodically) import JSON-AD files from a remote source.
 */
 
-use crate::{errors::AtomicResult, urls, Resource, Storelike};
+use crate::{errors::AtomicResult, storelike::Query, urls, Resource, Storelike};
 
 /// When an importer is shown, we list a bunch of Paramaters and a list of previously imported items.
 #[tracing::instrument(skip(store, query_params))]
@@ -31,8 +31,10 @@ pub fn construct_importer(
     }
 
     let parse_opts = crate::parse::ParseOpts {
-        for_agent: for_agent.map(|s| s.to_string()),
-        importer: Some(requested_subject),
+        // We sign the importer Commits with the default agent,
+        // not the one performing the import, because we don't have their private key.
+        for_agent: Some(store.get_default_agent()?),
+        importer: Some(requested_subject.clone()),
         create_commits: true,
         add: true,
     };
@@ -43,5 +45,26 @@ pub fn construct_importer(
 
     // TODO: generate list of imported resources
 
+    let q = Query {
+        property: Some(urls::PARENT.into()),
+        value: Some(requested_subject.into()),
+        limit: Some(10),
+        start_val: None,
+        end_val: None,
+        offset: 0,
+        sort_by: None,
+        sort_desc: false,
+        include_nested: true,
+        include_external: false,
+        for_agent: for_agent.map(|s| s.to_string()),
+    };
+    let results = store.query(&q)?;
+
+    resource.set_propval_unsafe(urls::ENDPOINT_RESULTS.into(), results.resources.into());
+    resource.set_propval(
+        urls::ENDPOINT_PARAMETERS.into(),
+        vec![urls::IMPORTER_JSON, urls::IMPORTER_URL].into(),
+        store,
+    )?;
     Ok(resource.clone())
 }