+++ lib/src/storelike.rs
@@ -8,7 +8,7 @@ use crate::{
     schema::{Class, Property},
     values::query_value_compare,
 };
-use crate::{errors::AtomicResult, parse::parse_json_ad_array};
+use crate::{errors::AtomicResult, parse::parse_json_ad_string};
 use crate::{mapping::Mapping, values::Value, Atom, Resource};
 
 // A path can return one of many things
@@ -209,7 +209,7 @@ pub trait Storelike: Sized {
 
     /// Imports a JSON-AD string, returns the amount of imported resources
     fn import(&self, string: &str, parse_opts: &crate::parse::ParseOpts) -> AtomicResult<usize> {
-        let vec = parse_json_ad_array(string, self, parse_opts)
+        let vec = parse_json_ad_string(string, self, parse_opts)
             .map_err(|e| format!("Unable to import JSON-AD. {}", e))?;
         let len = vec.len();
         Ok(len)